w3c-reports/12092014/gtk/runned-tests/webvtt/webvtt-api-for-browsers/vttcue-interface/getCueAsHTML-expected.txt



PASS VTTCue.getCueAsHTML(), creating the cue 
PASS VTTCue.getCueAsHTML(), <c> 
PASS VTTCue.getCueAsHTML(), <c.a.b> 
PASS VTTCue.getCueAsHTML(), <i> 
PASS VTTCue.getCueAsHTML(), <b> 
PASS VTTCue.getCueAsHTML(), <u> 
PASS VTTCue.getCueAsHTML(), <ruby> 
PASS VTTCue.getCueAsHTML(), <rt> 
PASS VTTCue.getCueAsHTML(), <v> 
PASS VTTCue.getCueAsHTML(), <v a b> 
PASS VTTCue.getCueAsHTML(), <00:00:00.500> 
FAIL VTTCue.getCueAsHTML(), x\0 assert_equals: data expected "x\0" but got "x"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/webvtt/webvtt-api-for-browsers/vttcue-interface/getCueAsHTML.html:89:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/webvtt/webvtt-api-for-browsers/vttcue-interface/getCueAsHTML.html:88:5)
